Final version of SW18AB V2.1.2

+++ SerialWombat18A_18B/SerialWombat18A_18B.X/swMath.c
@@ -32,9 +32,9 @@ uint16_t xyInterpolationU16(uint16_t x, uint16_t xl,uint16_t yl, uint16_t xh, ui
 	{
 		return (yh);
 	}
-	int32_t rise = yh- yl;
-	int32_t run = xh-xl;
-	int32_t val = x-xl;
+	int32_t rise = ((int32_t)yh)- yl;
+	int32_t run = ((int32_t)xh)-xl;
+	int32_t val = ((int32_t)x)-xl;
 	val *= rise;
 	val /= run;
 	val += yl;
